third_party/blink/web_tests/external/wpt/css/css-pseudo/placeholder-inherit-expected.txt

This is a testharness.js-based test.
[FAIL] ::placeholder should inherit from its originating element
  assert_equals: Check that ::placeholder is supported via color expected "rgb(0, 128, 0)" but got "rgb(0, 0, 0)"
Harness: the test ran to completion.
